+++ app/serializers/subscribe_serializer.rb
@@ -1,5 +1,5 @@
 class SubscribeSerializer < ActiveModel::Serializer
-  attributes :id, :status, :comment, :character
+  attributes :id, :status, :for_role, :comment, :character
 
   def character
     CharacterSubscriptionSerializer.new(object.character)